+++ NotifierRedirecter/Config.cs
@@ -1,7 +1,8 @@
-namespace NotifierRedirecter;
-
-public class Config
+namespace NotifierRedirecter
 {
-    public required string Token { get; set; }
-    public required ulong RedirectionChannel { get; set; }
+    public class Config
+    {
+        public required string Token { get; set; }
+        public required ulong RedirectionChannel { get; set; }
+    }
 }
